requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
python-dotenv==1.0.1
requests==2.32.3
email-validator==2.2.0
Flask-WTF==1.2.1
WTForms==3.1.2
gunicorn==21.2.0
cloudinary==1.41.0
sentry-sdk==1.45.0
Flask-Caching==2.3.0
redis==5.0.4
celery==5.4.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
stripe==8.9.0
pytest==7.4.0
eventlet==0.33.3
Flask-Limiter==2.8.1
algoliasearch==4.34.0

# Required for MySQL/PyMySQL sha256/caching_sha2_password auth
cryptography==41.0.4
